Use Arrays instead of Symbol for pool performance
See implementation with arrays: #307
See performance tests: #308

Closes: #300
Closes: #309
