+++ lib/poolify.js
@@ -1,12 +1,8 @@
 'use strict';
 
-const poolified = Symbol('poolified');
-
-const mixFlag = { [poolified]: true };
-
-const duplicate = (factory, n) => Array
-  .from({ length: n }, factory)
-  .map(instance => Object.assign(instance, mixFlag));
+const duplicate = (factory, n) => (
+  new Array(n).fill().map(() => factory())
+);
 
 const provide = callback => item => {
   setImmediate(() => {
@@ -17,10 +13,13 @@ const provide = callback => item => {
 const poolify = (factory, min, norm, max) => {
   let allocated = norm;
   const pool = (par) => {
-    if (par && par[poolified]) {
-      const delayed = pool.delayed.shift();
-      if (delayed) delayed(par);
-      else pool.items.push(par);
+    if (Array.isArray(par)) {
+      while (par.length) {
+        const item = par.shift();
+        const delayed = pool.delayed.shift();
+        if (delayed) delayed(item);
+        else pool.items.push(item);
+      }
       return;
     }
     if (pool.items.length < min && allocated < max) {
