+++ lib/poolify.symbol.js
@@ -1,8 +1,12 @@
 'use strict';
 
-const duplicate = (factory, n) => (
-  new Array(n).fill().map(() => factory())
-);
+const poolified = Symbol('poolified');
+
+const mixFlag = { [poolified]: true };
+
+const duplicate = (factory, n) => Array
+  .from({ length: n }, factory)
+  .map(instance => Object.assign(instance, mixFlag));
 
 const provide = callback => item => {
   setImmediate(() => {
@@ -13,13 +17,10 @@ const provide = callback => item => {
 const poolify = (factory, min, norm, max) => {
   let allocated = norm;
   const pool = (par) => {
-    if (Array.isArray(par)) {
-      while (par.length) {
-        const item = par.shift();
-        const delayed = pool.delayed.shift();
-        if (delayed) delayed(item);
-        else pool.items.push(item);
-      }
+    if (par && par[poolified]) {
+      const delayed = pool.delayed.shift();
+      if (delayed) delayed(par);
+      else pool.items.push(par);
       return;
     }
     if (pool.items.length < min && allocated < max) {
