+++ test/poolify-symbol
@@ -15,8 +15,8 @@ tap.test('poolify simple', (test) => {
     pool(item2 => {
       test.strictSame(pool.items.length, 13);
       test.strictSame(item2.length, 128);
-      pool([item1]);
-      pool([item2]);
+      pool(item1);
+      pool(item2);
       test.strictSame(pool.items.length, 15);
       test.end();
     });
@@ -32,7 +32,7 @@ tap.test('poolify loop', (test) => {
 
   for (let i = 0; i < 15; i++) {
     pool(item => {
-      pool([item]);
+      pool(item);
       if (i === 14) {
         test.strictSame(item.length, 128);
         test.end();
@@ -51,7 +51,7 @@ tap.test('poolify max', (test) => {
   for (let i = 0; i < 15; i++) {
     pool(item => {
       setTimeout(() => {
-        pool([item]);
+        pool(item);
         if (i === 14) {
           test.end();
         }
@@ -76,16 +76,16 @@ tap.test('poolify delayed order', (test) => {
         test.strictSame(pool.items.length, 0);
         test.strictSame(get3, false);
         get3 = true;
-        pool([item3]);
+        pool(item3);
       });
       pool(item4 => {
         test.strictSame(pool.items.length, 1);
         test.strictSame(get3, true);
-        pool([item4]);
+        pool(item4);
         test.end();
       });
-      pool([item1]);
-      pool([item2]);
+      pool(item1);
+      pool(item2);
     });
   });
 
@@ -101,8 +101,8 @@ tap.test('poolify functor', (test) => {
     test.strictSame(pool.items.length, 1);
     pool(item2 => {
       test.strictSame(pool.items.length, 0);
-      pool([item1]);
-      pool([item2]);
+      pool(item1);
+      pool(item2);
       test.strictSame(pool.items.length, 2);
       test.end();
     });
@@ -125,9 +125,9 @@ tap.test('poolify get sync', (test) => {
   const item4 = pool();
   test.strictSame(item4, undefined);
   test.strictSame(pool.items.length, 0);
-  pool([item1]);
-  pool([item2]);
-  pool([item3]);
+  pool(item1);
+  pool(item2);
+  pool(item3);
   test.strictSame(pool.items.length, 3);
   test.end();
